+++ Backend/IReceiver.cs
@@ -1,0 +1,13 @@
+﻿using Backend.Events;
+using Backend.Model;
+using System;
+
+namespace Backend {
+
+    public interface IReceiver {
+
+        event EventHandler<LogReceivedEventArgs> LogReceived;
+
+        void Initialize(Configuration configuration);
+    }
+}